requirements.txt
GitPython==3.1.26
wget==3.2
